# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Record Shop",
        "description": "An example application for a Record Shop selling Oldies But Goldies records.",
        "contact": {
            "email": "ismail.atkurt@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/records": {
            "get": {
                "tags": [
                    "records"
                ],
                "summary": "Get all records",
                "description": "Returns all records available",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    },
                    {
                        "name": "searchTerm",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of records"
                    }
                }
            },
            "post": {
                "tags": [
                    "records"
                ],
                "summary": "Create new record",
                "description": "Create new record with given data",
                "operationId": "create",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of the record",
                                        "required": [
                                            "true"
                                        ],
                                        "type": "string"
                                    },
                                    "genre": {
                                        "description": "Genre of the record",
                                        "required": [
                                            "false"
                                        ],
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Description of the record",
                                        "required": [
                                            "false"
                                        ],
                                        "type": "string"
                                    },
                                    "publishedAt": {
                                        "description": "Published date of the record",
                                        "required": [
                                            "false"
                                        ]
                                    },
                                    "artistId": {
                                        "description": "ID of related artist",
                                        "type": "int"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Down By The River",
                                    "artistId": 52,
                                    "genre": "Classic Rock",
                                    "description": "some record description",
                                    "publishedAt": "1966-03-26"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Returns Created Record"
                    }
                }
            }
        },
        "/records/{recordId}": {
            "get": {
                "tags": [
                    "records"
                ],
                "summary": "Get single record",
                "description": "Returns the record with given ID.",
                "operationId": "one",
                "parameters": [
                    {
                        "name": "recordId",
                        "in": "path",
                        "description": "The id of the record to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns Single Record"
                    }
                }
            },
            "put": {
                "tags": [
                    "records"
                ],
                "summary": "Update record",
                "description": "Update the record with given ID and provided post data.",
                "operationId": "update",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Updated name of the record",
                                        "type": "string"
                                    },
                                    "genre": {
                                        "description": "Genre of the record",
                                        "required": [
                                            "false"
                                        ],
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Description of the record",
                                        "required": [
                                            "false"
                                        ],
                                        "type": "string"
                                    },
                                    "publishedAt": {
                                        "description": "Published date of the record",
                                        "required": [
                                            "false"
                                        ]
                                    },
                                    "artistId": {
                                        "description": "ID of related artist",
                                        "type": "int"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Down By The River 2",
                                    "artistId": 52,
                                    "genre": "Pop Music",
                                    "description": "some updated record description",
                                    "publishedAt": "1976-04-27"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Returns Updated Record"
                    }
                }
            },
            "delete": {
                "tags": [
                    "records"
                ],
                "summary": "Deletes a record",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "recordId",
                        "in": "path",
                        "description": "Id of the record to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Validation error message"
                    },
                    "404": {
                        "description": "Record not found"
                    }
                }
            }
        }
    }
}